增加BindingResult的处理逻辑

+++ spring-cloud-consumer-book/src/main/java/com/wujunshen/web/controller/BookConsumerController.java
@@ -13,6 +13,7 @@ import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.http.MediaType;
 import org.springframework.util.ObjectUtils;
+import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -53,14 +54,19 @@ public class BookConsumerController extends BaseController {
             @ApiResponse(code = 403, message = "Forbidden"),
             @ApiResponse(code = 404, message = "Not Found"),
             @ApiResponse(code = 500, message = "Failure")})
-    public BaseResponse saveBook(@Validated @ApiParam(value = "添加的某本书籍信息", required = true) @RequestBody Book book) {
+    public BaseResponse saveBook(@Validated @ApiParam(value = "添加的某本书籍信息", required = true) @RequestBody Book book, BindingResult bindingResult) {
+        BaseResponse baseResponse = getValidatedResult(bindingResult);
+        if (baseResponse != null) {
+            return baseResponse;
+        }
+
         String jwtToken = stringRedisTemplate.opsForValue().get(Constants.BEARER);
 
         ValueOperations<String, LoginParameter> operations = redisTemplate.opsForValue();
         LoginParameter loginParameter = operations.get(Constants.LOGIN_PARAMETER);
 
         if (!ObjectUtils.isEmpty(jwtToken)) {
-            BaseResponse baseResponse = bookConsumerService.saveBook(Constants.BEARER + " " + jwtToken, book);
+            baseResponse = bookConsumerService.saveBook(Constants.BEARER + " " + jwtToken, book);
 
             if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
                 return baseResponse;
@@ -83,14 +89,19 @@ public class BookConsumerController extends BaseController {
             @ApiResponse(code = 403, message = "Forbidden"),
             @ApiResponse(code = 404, message = "Not Found"),
             @ApiResponse(code = 500, message = "Failure")})
-    public BaseResponse getBooks() {
+    public BaseResponse getBooks(BindingResult bindingResult) {
+        BaseResponse baseResponse = getValidatedResult(bindingResult);
+        if (baseResponse != null) {
+            return baseResponse;
+        }
+
         String jwtToken = stringRedisTemplate.opsForValue().get(Constants.BEARER);
 
         ValueOperations<String, LoginParameter> operations = redisTemplate.opsForValue();
         LoginParameter loginParameter = operations.get(Constants.LOGIN_PARAMETER);
 
         if (!ObjectUtils.isEmpty(jwtToken)) {
-            BaseResponse baseResponse = bookConsumerService.getBooks(Constants.BEARER + " " + jwtToken);
+            baseResponse = bookConsumerService.getBooks(Constants.BEARER + " " + jwtToken);
 
             if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
                 return baseResponse;
@@ -110,7 +121,12 @@ public class BookConsumerController extends BaseController {
             @ApiResponse(code = 403, message = "Forbidden"),
             @ApiResponse(code = 404, message = "Not Found"),
             @ApiResponse(code = 500, message = "Failure")})
-    public BaseResponse getBook(@ApiParam(value = "书籍ID", required = true) @PathVariable Integer bookId) {
+    public BaseResponse getBook(@ApiParam(value = "书籍ID", required = true) @PathVariable Integer bookId, BindingResult bindingResult) {
+        BaseResponse baseResponse = getValidatedResult(bindingResult);
+        if (baseResponse != null) {
+            return baseResponse;
+        }
+
         //1.从redis取token，没有取到token就调用getBaseResponse方法，有就执行getBook方法
         //2.执行getBook方法后，如果发现token无效，则也调用getBaseResponse方法，有效就返回getBook方法的响应消息
         String jwtToken = stringRedisTemplate.opsForValue().get(Constants.BEARER);
@@ -119,7 +135,7 @@ public class BookConsumerController extends BaseController {
         LoginParameter loginParameter = operations.get(Constants.LOGIN_PARAMETER);
 
         if (!ObjectUtils.isEmpty(jwtToken)) {
-            BaseResponse baseResponse = bookConsumerService.getBook(Constants.BEARER + " " + jwtToken, bookId);
+            baseResponse = bookConsumerService.getBook(Constants.BEARER + " " + jwtToken, bookId);
 
             if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
                 return baseResponse;
@@ -139,14 +155,19 @@ public class BookConsumerController extends BaseController {
             @ApiResponse(code = 403, message = "Forbidden"),
             @ApiResponse(code = 404, message = "Not Found"),
             @ApiResponse(code = 500, message = "Failure")})
-    public BaseResponse updateBook(@ApiParam(value = "要更新的某本书籍ID", required = true) @PathVariable("bookId") Integer bookId, @ApiParam(value = "要更新的某本书籍信息", required = true) @RequestBody Book book) {
+    public BaseResponse updateBook(@ApiParam(value = "要更新的某本书籍ID", required = true) @PathVariable("bookId") Integer bookId, @ApiParam(value = "要更新的某本书籍信息", required = true) @RequestBody Book book, BindingResult bindingResult) {
+        BaseResponse baseResponse = getValidatedResult(bindingResult);
+        if (baseResponse != null) {
+            return baseResponse;
+        }
+
         String jwtToken = stringRedisTemplate.opsForValue().get(Constants.BEARER);
 
         ValueOperations<String, LoginParameter> operations = redisTemplate.opsForValue();
         LoginParameter loginParameter = operations.get(Constants.LOGIN_PARAMETER);
 
         if (!ObjectUtils.isEmpty(jwtToken)) {
-            BaseResponse baseResponse = bookConsumerService.updateBook(Constants.BEARER + " " + jwtToken, bookId, book);
+            baseResponse = bookConsumerService.updateBook(Constants.BEARER + " " + jwtToken, bookId, book);
 
             if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
                 return baseResponse;
@@ -166,14 +187,19 @@ public class BookConsumerController extends BaseController {
             @ApiResponse(code = 403, message = "Forbidden"),
             @ApiResponse(code = 404, message = "Not Found"),
             @ApiResponse(code = 500, message = "Failure")})
-    public BaseResponse deleteBook(@ApiParam(value = "要删除的某本书籍ID", required = true) @PathVariable("bookId") Integer bookId) {
+    public BaseResponse deleteBook(@ApiParam(value = "要删除的某本书籍ID", required = true) @PathVariable("bookId") Integer bookId, BindingResult bindingResult) {
+        BaseResponse baseResponse = getValidatedResult(bindingResult);
+        if (baseResponse != null) {
+            return baseResponse;
+        }
+
         String jwtToken = stringRedisTemplate.opsForValue().get(Constants.BEARER);
 
         ValueOperations<String, LoginParameter> operations = redisTemplate.opsForValue();
         LoginParameter loginParameter = operations.get(Constants.LOGIN_PARAMETER);
 
         if (!ObjectUtils.isEmpty(jwtToken)) {
-            BaseResponse baseResponse = bookConsumerService.deleteBook(Constants.BEARER + " " + jwtToken, bookId);
+            baseResponse = bookConsumerService.deleteBook(Constants.BEARER + " " + jwtToken, bookId);
 
             if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
                 return baseResponse;
@@ -181,7 +207,6 @@ public class BookConsumerController extends BaseController {
         }
         return deleteBook(bookId, loginParameter);
     }
-
 
     /**
      * 调用获取token的方法，成功就直接请求getBook方法，否则就返回出错的响应消息
@@ -195,12 +220,12 @@ public class BookConsumerController extends BaseController {
 
         if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
             Map tokenMap = (HashMap) baseResponse.getData();
-            log.info("token_type is: {}", tokenMap.get("token_type"));
-            log.info("access_token is: {}", tokenMap.get("access_token"));
-            log.info("expires_in is: {}", tokenMap.get("expires_in"));
+            log.info(Constants.TOKEN_TYPE + " {}", tokenMap.get(Constants.TOKEN_TYPE));
+            log.info(Constants.ACCESS_TOKEN + " {}", tokenMap.get(Constants.ACCESS_TOKEN));
+            log.info(Constants.EXPIRES_IN + " {}", tokenMap.get(Constants.EXPIRES_IN));
 
-            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get("access_token"), ((Integer) tokenMap.get("expires_in")).longValue(), TimeUnit.SECONDS);
-            return bookConsumerService.getBook(Constants.BEARER + " " + tokenMap.get("access_token"), bookId);
+            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get(Constants.ACCESS_TOKEN), ((Integer) tokenMap.get(Constants.EXPIRES_IN)).longValue(), TimeUnit.SECONDS);
+            return bookConsumerService.getBook(Constants.BEARER + " " + tokenMap.get(Constants.ACCESS_TOKEN), bookId);
         }
 
         return baseResponse;
@@ -218,12 +243,12 @@ public class BookConsumerController extends BaseController {
 
         if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
             Map tokenMap = (HashMap) baseResponse.getData();
-            log.info("token_type is: {}", tokenMap.get("token_type"));
-            log.info("access_token is: {}", tokenMap.get("access_token"));
-            log.info("expires_in is: {}", tokenMap.get("expires_in"));
+            log.info(Constants.TOKEN_TYPE + " {}", tokenMap.get(Constants.TOKEN_TYPE));
+            log.info(Constants.ACCESS_TOKEN + " {}", tokenMap.get(Constants.ACCESS_TOKEN));
+            log.info(Constants.EXPIRES_IN + " {}", tokenMap.get(Constants.EXPIRES_IN));
 
-            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get("access_token"), ((Integer) tokenMap.get("expires_in")).longValue(), TimeUnit.SECONDS);
-            return bookConsumerService.saveBook(Constants.BEARER + " " + tokenMap.get("access_token"), book);
+            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get(Constants.ACCESS_TOKEN), ((Integer) tokenMap.get(Constants.EXPIRES_IN)).longValue(), TimeUnit.SECONDS);
+            return bookConsumerService.saveBook(Constants.BEARER + " " + tokenMap.get(Constants.ACCESS_TOKEN), book);
         }
 
         return baseResponse;
@@ -240,12 +265,12 @@ public class BookConsumerController extends BaseController {
 
         if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
             Map tokenMap = (HashMap) baseResponse.getData();
-            log.info("token_type is: {}", tokenMap.get("token_type"));
-            log.info("access_token is: {}", tokenMap.get("access_token"));
-            log.info("expires_in is: {}", tokenMap.get("expires_in"));
+            log.info(Constants.TOKEN_TYPE + " {}", tokenMap.get(Constants.TOKEN_TYPE));
+            log.info(Constants.ACCESS_TOKEN + " {}", tokenMap.get(Constants.ACCESS_TOKEN));
+            log.info(Constants.EXPIRES_IN + " {}", tokenMap.get(Constants.EXPIRES_IN));
 
-            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get("access_token"), ((Integer) tokenMap.get("expires_in")).longValue(), TimeUnit.SECONDS);
-            return bookConsumerService.getBooks(Constants.BEARER + " " + tokenMap.get("access_token"));
+            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get(Constants.ACCESS_TOKEN), ((Integer) tokenMap.get(Constants.EXPIRES_IN)).longValue(), TimeUnit.SECONDS);
+            return bookConsumerService.getBooks(Constants.BEARER + " " + tokenMap.get(Constants.ACCESS_TOKEN));
         }
 
         return baseResponse;
@@ -262,12 +287,12 @@ public class BookConsumerController extends BaseController {
 
         if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
             Map tokenMap = (HashMap) baseResponse.getData();
-            log.info("token_type is: {}", tokenMap.get("token_type"));
-            log.info("access_token is: {}", tokenMap.get("access_token"));
-            log.info("expires_in is: {}", tokenMap.get("expires_in"));
+            log.info(Constants.TOKEN_TYPE + " {}", tokenMap.get(Constants.TOKEN_TYPE));
+            log.info(Constants.ACCESS_TOKEN + " {}", tokenMap.get(Constants.ACCESS_TOKEN));
+            log.info(Constants.EXPIRES_IN + " {}", tokenMap.get(Constants.EXPIRES_IN));
 
-            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get("access_token"), ((Integer) tokenMap.get("expires_in")).longValue(), TimeUnit.SECONDS);
-            return bookConsumerService.updateBook(Constants.BEARER + " " + tokenMap.get("access_token"), bookId, book);
+            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get(Constants.ACCESS_TOKEN), ((Integer) tokenMap.get(Constants.EXPIRES_IN)).longValue(), TimeUnit.SECONDS);
+            return bookConsumerService.updateBook(Constants.BEARER + " " + tokenMap.get(Constants.ACCESS_TOKEN), bookId, book);
         }
 
         return baseResponse;
@@ -284,12 +309,12 @@ public class BookConsumerController extends BaseController {
 
         if (baseResponse.getCode() == ResponseStatus.OK.getCode()) {
             Map tokenMap = (HashMap) baseResponse.getData();
-            log.info("token_type is: {}", tokenMap.get("token_type"));
-            log.info("access_token is: {}", tokenMap.get("access_token"));
-            log.info("expires_in is: {}", tokenMap.get("expires_in"));
+            log.info(Constants.TOKEN_TYPE + " {}", tokenMap.get(Constants.TOKEN_TYPE));
+            log.info(Constants.ACCESS_TOKEN + " {}", tokenMap.get(Constants.ACCESS_TOKEN));
+            log.info(Constants.EXPIRES_IN + " {}", tokenMap.get(Constants.EXPIRES_IN));
 
-            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get("access_token"), ((Integer) tokenMap.get("expires_in")).longValue(), TimeUnit.SECONDS);
-            return bookConsumerService.deleteBook(Constants.BEARER + " " + tokenMap.get("access_token"), bookId);
+            stringRedisTemplate.opsForValue().set(Constants.BEARER, (String) tokenMap.get(Constants.ACCESS_TOKEN), ((Integer) tokenMap.get(Constants.EXPIRES_IN)).longValue(), TimeUnit.SECONDS);
+            return bookConsumerService.deleteBook(Constants.BEARER + " " + tokenMap.get(Constants.ACCESS_TOKEN), bookId);
         }
 
         return baseResponse;
